+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Add `Store` contract compatible with solidity `^0.8.0`. (#138)
 - Add Boost module. (#149 #151 #153)
     - Add interfaces. (#156)   
+- Deploy Nominal and Boost modules with Hardhat. (#)
 
 ## v5.4.2
 - Use latest Nethermind base image `nethermindeth/nethermind:iexec`